fix android user signup
